Move common candidate applications logic into helper
These methods will be used by a couple of email services so make them
easily available.
